test/requirements.txt
parameterized==0.9.0
python-dotenv==1.0.0
requests_mock==1.11.0